+++ src/SFA.DAS.Tools.Support.Infrastructure/Services/EmployerCommitmentsService.cs
@@ -13,7 +13,6 @@ namespace SFA.DAS.Tools.Support.Infrastructure.Services
     public interface IEmployerCommitmentsService
     {
         Task<StopApprenticeshipResult> StopApprenticeship(long employerAccountId, long apprenticeshipId, string UserId, DateTime stopDate);
-        Task<ApprenticeshipSummaryResult> GetApprenticeshipSummary(long apprenticeshipId, long employerAccountId);
         Task<SearchApprenticeshipsResult> SearchApprenticeships(string courseName, string employerName, string providerName, string searchTerm, DateTime? startDate, DateTime? endDate);
         Task<GetApprenticeshipResult> GetApprenticeship(long id);
     }
@@ -59,37 +58,6 @@ namespace SFA.DAS.Tools.Support.Infrastructure.Services
             {
                 _logger.LogError(e, "Failure to stop the apprenticeship.");
                 return new StopApprenticeshipResult
-                {
-                    ErrorMessage = e.Message
-                };
-            }
-        }
-
-        public async Task<ApprenticeshipSummaryResult> GetApprenticeshipSummary(long apprenticeshipId, long employerAccountId)
-        {
-            try
-            {
-                if (employerAccountId <= 0)
-                {
-                    throw new ArgumentException("employerAccountId must be greater than 0", "employerAccountId");
-                }
-
-                if (apprenticeshipId <= 0)
-                {
-                    throw new ArgumentException("apprenticeshipId must be greater than 0", "apprenticeshipId");
-                }
-
-                // switch for get employer apprenticeship
-                //var result = await _commitmentApi.GetEmployerApprenticeship(employerAccountId, apprenticeshipId);
-
-                //return _mapper.Map<ApprenticeshipSummaryResult>(result);
-                await Task.FromResult(0);
-                return new ApprenticeshipSummaryResult();
-            }
-            catch (Exception e)
-            {
-                _logger.LogError(e, "Failed to fetch the apprenticeship.");
-                return new ApprenticeshipSummaryResult
                 {
                     ErrorMessage = e.Message
                 };
