+++ src/SFA.DAS.Tools.Support.Web/Views/Approvals/SearchApprenticeships.cshtml
@@ -178,8 +178,8 @@
             $('#submitSearchFormButton').on('click', function (evt) {
                 evt.preventDefault();
                 $.post('', $('form').serialize(), function (data) {
-                    if (data.ErrorTitle != null) {
-                        alert(data.ErrorMessage)
+                    if (data.errorTitle != null) {
+                        alert(data.errorTitle + "\r\n\r\n" + data.errorMessage)
                     } else {
                         $('html,body').animate({ scrollTop: $('#apprenticeshipResultsTable').offset().top }, 'slow');
                         table.bootstrapTable('load', data);
@@ -189,7 +189,7 @@
         });
 
         function createStopLink(value, row, index) {
-            return '<a class="detail" href=' + ($('#baseUrl').val() + "/?id=" + row.id) + ' title ="Stop">Stop</a>';
+            return '<a class="detail" href=' + ($('#baseUrl').val() + "?id=" + row.id) + ' title ="Stop">Stop</a>';
         };
     </script>
 }
