--- src/SFA.DAS.Tools.Support.Web/Views/Approvals/StopApprenticeshipConfirmation.cshtml
@@ -1,63 +1,0 @@
-﻿@{
-    ViewData["Title"] = "Support Portal";
-    @model StopApprenticeshipConfirmationViewModel
-    @using SFA.DAS.Tools.Support.Web.Configuration
-}
-
-@if (Model != null && Model.SubmittedSuccessfully)
-{
-    <div class="govuk-grid-row">
-        <div class="govuk-grid-column-two-thirds">
-            <div class="govuk-panel govuk-panel--confirmation">
-                <h1 class="govuk-panel__title">
-                    Apprenticeship has been stopped.
-                </h1>
-                <div class="govuk-panel__body">
-                    Apprenticeship with Id @Model.ApprenticeshipId, and EmployerAccountId @Model.EmployerAccountId has been marked to be stopped on @Model.EnteredStopDate.ToShortDateString().
-                </div>
-            </div>
-        </div>
-    </div>
-
-}
-else
-{
-    <div class="govuk-grid-row">
-        <div class="govuk-grid-column-two-thirds">
-            <h1 class="govuk-heading-xl">
-                Stop an Apprenticeship Confirmation
-            </h1>
-            <p> This is the Apprenticeship being stopped. Is this correct?</p>
-            <ul class="govuk-list">
-                <li>Apprentice Name: @(Model.ApprenticeshipName ?? "-")</li>
-                <li>Provider Name: @(Model.ProviderName ?? "-")</li>
-                <li>Endpoint Assessors Name: @(Model.EndpointAssessorName ?? "-")</li>
-                <li>Completion Date: @(Model.CompletionDate?.ToShortDateString() ?? "-")</li>
-                <li>Stop Date: @(Model.StopDate?.ToShortDateString() ?? "-")</li>
-                <li>Payment Status: @(Model.PaymentStatus)</li>
-                <li>ULN: @(Model.ULN ?? "-")</li>
-                <li>Start Date: @(Model.StartDate?.ToShortDateString() ?? "-")</li>
-                <li>End Date: @(Model.EndDate?.ToShortDateString() ?? "-")</li>
-                <li>Pause Date: @(Model.PauseDate?.ToShortDateString() ?? "-")</li>
-            </ul>
-            <div asp-validation-summary="ModelOnly" class="govuk-error-message"></div>
-            <form asp-route="@ApprovalsRouteNames.StopApprenticeshipConfirmation" method="post">
-                <div class="govuk-form-group">
-                    @Html.HiddenFor(s => s.EmployerAccountId)
-                    @Html.HiddenFor(s => s.ApprenticeshipId)
-                    @Html.HiddenFor(s => s.EnteredStopDate)
-                    <button class="govuk-button" data-module="govuk-button" name="submit" value="cancel">
-                        Back
-                    </button>
-                    <button class="govuk-button" data-module="govuk-button" name="submit" value="submit">
-                        Confirm
-                    </button>
-                </div>
-            </form>
-
-        </div>
-    </div>
-}
-@section scripts {
-
-}
